Create optimizedAlgo.java
I had tried to write the logic: sort the item price from low to high, then deduct the money balance 900 bucks from the lowest price item until no item that the balance can be buy, but I realised this logic not able to achieve maximize the use of money. For instance, when the amount of 900 bucks are changed to 800 bucks, the best case is buy item with 450 and 350 bucks, where remaining will be zero, but my logic will buy items with 300 and 200 bucks due to the early items sorted.